Stupid bullshit to make systray work with xapps
Xapps provides the Host and Watcher for StatusNotiferItem's (SNI) on a
couple DE's like Cinnamon and Xfce. It seems to not call the ContextMenu
method on right click and it does not support dbus menu's. This means
that the only useful SNI Method we can use is Activate (left click). So
as a workaround I just put the menu popup which I ideally would only
link to right click to all callbacks.

I will submit an issue about this at the xapps repo asap.